+++ src/constants/index.js
@@ -1,4 +1,4 @@
-const VERSION = '1.5.0'
+const VERSION = '1.5.1'
 const BLOCK_ROWS = 50
 const CLUSTER_BLOCKS = 4