+++ src/vue/MultipleSelect.vue
@@ -99,6 +99,16 @@ export default {
     }
   },
 
+  beforeUpdate () {
+    if (this.slotDefault || this.slotDefault !== this.$slots.default) {
+      this.slotDefault = this.$slots.default
+      this.$nextTick(() => {
+        this._refresh()
+        this._initSelectValue()
+      })
+    }
+  },
+
   mounted () {
     this._refresh()