+++ vue-examples/src/examples/Locale.vue
@@ -78,10 +78,6 @@
 </template>
 
 <script>
-import 'multiple-select/dist/multiple-select-locale-all'
-import $ from 'jquery'
-$.extend($.fn.multipleSelect.defaults, $.fn.multipleSelect.locales['en-US'])
-
 export default {
   data () {
     return {